cleanup(perception): Nice cleanup of traffic object indexes by adding abstract methods to be called on addition/removal/updates of indexes
