+++ fed/mosaic-application/src/main/java/org/eclipse/mosaic/fed/application/ambassador/simulation/perception/index/providers/TrafficLightMap.java
@@ -15,15 +15,10 @@
 
 package org.eclipse.mosaic.fed.application.ambassador.simulation.perception.index.providers;
 
-import org.eclipse.mosaic.fed.application.ambassador.SimulationKernel;
 import org.eclipse.mosaic.fed.application.ambassador.simulation.perception.PerceptionModel;
 import org.eclipse.mosaic.fed.application.ambassador.simulation.perception.index.objects.TrafficLightObject;
-import org.eclipse.mosaic.lib.objects.trafficlight.TrafficLightGroup;
-import org.eclipse.mosaic.lib.objects.trafficlight.TrafficLightGroupInfo;
-import org.eclipse.mosaic.lib.objects.trafficlight.TrafficLightState;
 
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 public class TrafficLightMap extends TrafficLightIndex {
@@ -41,40 +36,8 @@ public class TrafficLightMap extends TrafficLightIndex {
     }
 
     @Override
-    public void addTrafficLight(TrafficLightGroup trafficLightGroup) {
-        String trafficLightGroupId = trafficLightGroup.getGroupId();
-        trafficLightGroup.getTrafficLights().forEach(
-                (trafficLight) -> {
-                    String trafficLightId = calculateTrafficLightId(trafficLightGroupId, trafficLight.getId());
-                    if (SimulationKernel.SimulationKernel.getCentralPerceptionComponent().getScenarioBounds()
-                            .contains(trafficLight.getPosition().toCartesian())) { // check if inside bounding area
-                        indexedTrafficLights.computeIfAbsent(trafficLightId, TrafficLightObject::new)
-                                .setTrafficLightGroupId(trafficLightGroupId)
-                                .setPosition(trafficLight.getPosition().toCartesian())
-                                .setIncomingLane(trafficLight.getIncomingLane())
-                                .setOutgoingLane(trafficLight.getOutgoingLane())
-                                .setTrafficLightState(trafficLight.getCurrentState());
-                    }
-                }
-        );
+    public void onTrafficLightsUpdate() {
+        // nothing to do on update
     }
 
-    @Override
-    public void updateTrafficLights(Map<String, TrafficLightGroupInfo> trafficLightGroupsToUpdate) {
-        trafficLightGroupsToUpdate.forEach(
-                (trafficLightGroupId, trafficLightGroupInfo) -> {
-                    List<TrafficLightState> trafficLightStates = trafficLightGroupInfo.getCurrentState();
-                    for (int i = 0; i < trafficLightStates.size(); i++) {
-                        String trafficLightId = calculateTrafficLightId(trafficLightGroupId, i);
-                        final TrafficLightState trafficLightState = trafficLightStates.get(i);
-                        indexedTrafficLights.computeIfPresent(trafficLightId, (id, trafficLightObject)
-                                -> trafficLightObject.setTrafficLightState(trafficLightState));
-                    }
-                }
-        );
-    }
-
-    private String calculateTrafficLightId(String trafficLightGroupId, int trafficLightIndex) {
-        return trafficLightGroupId + "_" + trafficLightIndex;
-    }
 }
