+++ fed/mosaic-application/src/main/java/org/eclipse/mosaic/fed/application/ambassador/simulation/perception/index/providers/TrafficLightTree.java
@@ -15,13 +15,9 @@
 
 package org.eclipse.mosaic.fed.application.ambassador.simulation.perception.index.providers;
 
-import org.eclipse.mosaic.fed.application.ambassador.SimulationKernel;
 import org.eclipse.mosaic.fed.application.ambassador.simulation.perception.PerceptionModel;
 import org.eclipse.mosaic.fed.application.ambassador.simulation.perception.index.objects.SpatialObjectAdapter;
 import org.eclipse.mosaic.fed.application.ambassador.simulation.perception.index.objects.TrafficLightObject;
-import org.eclipse.mosaic.lib.objects.trafficlight.TrafficLightGroup;
-import org.eclipse.mosaic.lib.objects.trafficlight.TrafficLightGroupInfo;
-import org.eclipse.mosaic.lib.objects.trafficlight.TrafficLightState;
 import org.eclipse.mosaic.lib.spatial.KdTree;
 import org.eclipse.mosaic.lib.spatial.SpatialTreeTraverser;
 
@@ -29,7 +25,6 @@ import com.google.gson.annotations.Expose;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 /**
@@ -46,7 +41,7 @@ public class TrafficLightTree extends TrafficLightIndex {
 
     @Override
     public void initialize() {
-        // nothing to initialize
+        // initialization at first update
     }
 
     @Override
@@ -58,50 +53,16 @@ public class TrafficLightTree extends TrafficLightIndex {
     }
 
     @Override
-    public void addTrafficLight(TrafficLightGroup trafficLightGroup) {
-        String trafficLightGroupId = trafficLightGroup.getGroupId();
-        trafficLightGroup.getTrafficLights().forEach(
-                (trafficLight) -> {
-                    String trafficLightId = calculateTrafficLightId(trafficLightGroupId, trafficLight.getId());
-                    if (SimulationKernel.SimulationKernel.getCentralPerceptionComponent().getScenarioBounds()
-                            .contains(trafficLight.getPosition().toCartesian())) {
-                        indexedTrafficLights.computeIfAbsent(trafficLightId, TrafficLightObject::new)
-                                .setTrafficLightGroupId(trafficLightGroupId)
-                                .setPosition(trafficLight.getPosition().toCartesian())
-                                .setIncomingLane(trafficLight.getIncomingLane())
-                                .setOutgoingLane(trafficLight.getOutgoingLane())
-                                .setTrafficLightState(trafficLight.getCurrentState());
-                    }
-                }
-        );
-    }
-
-    @Override
-    public void updateTrafficLights(Map<String, TrafficLightGroupInfo> trafficLightGroupsToUpdate) {
-        if (trafficLightTree == null) {
+    public void onTrafficLightsUpdate() {
+        if (trafficLightTree == null) { // initialize before first update is called
             List<TrafficLightObject> allTrafficLights = new ArrayList<>(indexedTrafficLights.values());
             trafficLightTree = new KdTree<>(new SpatialObjectAdapter<>(), allTrafficLights, bucketSize);
             treeTraverser = new SpatialTreeTraverser.InRadius<>();
         }
-        trafficLightGroupsToUpdate.forEach(
-                (trafficLightGroupId, trafficLightGroupInfo) -> {
-                    List<TrafficLightState> trafficLightStates = trafficLightGroupInfo.getCurrentState();
-                    for (int i = 0; i < trafficLightStates.size(); i++) { // check if inside bounding area
-                        String trafficLightId = calculateTrafficLightId(trafficLightGroupId, i);
-                        final TrafficLightState trafficLightState = trafficLightStates.get(i);
-                        indexedTrafficLights.computeIfPresent(trafficLightId, (id, trafficLightObject)
-                                -> trafficLightObject.setTrafficLightState(trafficLightState));
-                    }
-                }
-        );
     }
 
     @Override
     public int getNumberOfTrafficLights() {
         return trafficLightTree.getRoot().size();
-    }
-
-    private String calculateTrafficLightId(String trafficLightGroupId, int trafficLightIndex) {
-        return trafficLightGroupId + "_" + trafficLightIndex;
     }
 }
